+++ CHANGELOG.md
@@ -16,6 +16,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Changed
 - Make `-t` optional, default to `file-input`
 - Explicitly set python version to `3.10`
+- Split validation and checksum generation into different CLI commands
+- Re-organize functions
 ### Fixed
 
 ---
